+++ src/bin/uair/timer.rs
@@ -4,17 +4,19 @@ use async_io::Timer;
 use crate::Error;
 use crate::app::Event;
 use crate::session::Session;
+use crate::socket::BlockingStream;
 
 pub struct UairTimer {
 	interval: Duration,
+	streams: Vec<BlockingStream>,
 }
 
 impl UairTimer {
 	pub fn new(interval: Duration) -> Self {
-		UairTimer { interval }
+		UairTimer { interval, streams: Vec::new() }
 	}
 
-	pub async fn start(&self, session: &Session, start: Instant, dest: Instant) -> Result<Event, Error> {
+	pub async fn start(&mut self, session: &Session, start: Instant, dest: Instant) -> Result<Event, Error> {
 		let mut stdout = io::stdout();
 
 		let duration = dest - start;
@@ -23,11 +25,17 @@ impl UairTimer {
 
 		while end <= dest {
 			Timer::at(end).await;
-			write!(stdout, "{}", session.display::<true>(dest - end))?;
+			let time = session.display::<true>(dest - end).to_string() + "\0";
+			write!(stdout, "{}", time)?;
+			self.streams.retain_mut(|stream| stream.write(time.as_bytes()).is_ok());
 			stdout.flush()?;
 			end += self.interval;
 		}
 
 		Ok(Event::Finished)
+	}
+
+	pub fn add_stream(&mut self, stream: BlockingStream) {
+		self.streams.push(stream);
 	}
 }